clients/python/pydantic-ai/pyproject.toml
[project]
name = "ai-search-pydantic-mcp-client"
version = "0.1.0"
description = "An MCP Client for Azure AI Search Service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp==1.6.0",
    "openai>=1.77.0",
    "pydantic>=2.11.4",
    "pydantic-ai>=0.1.10",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
